Refactor scene description presentation
